Add React Router homepage with link to Order component
