+++ package.json
@@ -9,6 +9,8 @@
     "axios": "^0.20.0",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-router": "^5.2.0",
+    "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
     "yup": "^0.29.3"
   },
@@ -17,7 +19,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "cypress":"cypress open"
+    "cypress": "cypress open"
   },
   "eslintConfig": {
     "extends": "react-app"
