+++ src/App.js
@@ -1,117 +1,47 @@
-import React, { useState, useEffect } from 'react'
-import Pizza from './Pizza'
-import PizzaForm from './PizzaForm'
-import formSchema from './formSchema'
-import axios from 'axios'
-import * as yup from 'yup'
-
-const initialFormValues = {
-  name: '',
-  size: '',
-  toppings: {
-    hiking: false,
-    reading: false,
-    coding: false,
-  },
-  instructions: '',
-}
-const initialFormErrors = {
-  name: '',
-  size: '',
-}
-const initialPizzas = []
-const initialDisabled = true
-
+import React from "react";
+import {
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link
+} from "react-router-dom";
+import Order from './Order'
 
 export default function App() {
-  const [pizzas, setPizzas] = useState(initialPizzas)
-  const [formValues, setFormValues] = useState(initialFormValues)
-  const [formErrors, setFormErrors] = useState(initialFormErrors)
-  const [disabled, setDisabled] = useState(initialDisabled)
-
-  const postNewPizza = newPizza => {
-    axios.post('https://reqres.in/api/pizzas', newPizza)
-      .then(res => {
-        setPizzas([...pizzas, res.data])
-      })
-      .catch(err => {
-        return err
-      })
-      .finally(() => {
-        setFormValues(initialFormValues)
-      })
-  }
-
-  const inputChange = (name, value) => {
-    yup
-      .reach(formSchema, name)
-      .validate(value)
-      .then(valid => {
-        setFormErrors({
-          ...formErrors,
-          [name]: "",
-        })
-      })
-      .catch(err => {
-        setFormErrors({
-          ...formErrors,
-          [name]: err.errors[0],
-        })
-      })
-
-    setFormValues({
-      ...formValues,
-      [name]: value
-    })
-  }
-
-  const checkboxChange = (name, isChecked) => {
-    setFormValues({
-      ...formValues,
-      toppings: {
-        ...formValues.toppings,
-        [name]: isChecked,
-      }
-    })
-  }
-
-  const submit = () => {
-    const newPizza = {
-      name: formValues.name.trim(),
-      size: formValues.size,
-      toppings: Object.keys(formValues.toppings).filter(topping => formValues.toppings[topping]),
-      instructions: formValues.instructions.trim(),
-    }
-    postNewPizza(newPizza)
-  }
-
-  useEffect(() => {
-    formSchema.isValid(formValues)
-      .then(valid => {
-        setDisabled(!valid);
-      })
-  }, [formValues])
-
   return (
-    <div className='container'>
-      <header><h1>Lambda Eats</h1></header>
+    <Router>
+      <div>
+        <nav>
+          <ul>
+            <li>
+              <Link to="/">Home</Link>
+            </li>
+            <li>
+              <Link to="/order">Order</Link>
+            </li>
+          </ul>
+        </nav>
+        <Switch>
+          <Route path="/order">
+            <Order />
+          </Route>
+          <Route path="/">
+            <Home />
+          </Route>
+        </Switch>
+      </div>
+    </Router>
+  );
+}
 
-      <PizzaForm
-        values={formValues}
-        inputChange={inputChange}
-        checkboxChange={checkboxChange}
-        submit={submit}
-        disabled={disabled}
-        errors={formErrors}
-      />
+function Home() {
+  return <h2>Home</h2>;
+}
 
-      {
-        pizzas.map(pizza => {
-          return (
-            <Pizza key={pizza.id} details={pizza} />
-          )
-        })
-      }
-    </div>
-  )
+function About() {
+  return <h2>About</h2>;
+}
+
+function Users() {
+  return <h2>Users</h2>;
 }
